Refactor penalty group retrieval into dedicated service
